requirements.txt
streamlit
google-genai
pandas
plotly
Pillow